requirements.txt
langgraph 
langchain 
langchain-openai 
langchain-community 
duckduckgo-search 
python-dotenv
ddgs
fastapi 
uvicorn
opentelemetry-api 
opentelemetry-sdk
opentelemetry-exporter-jaeger
opentelemetry-exporter-otlp